1MB._RAM
A bloat file that equals exactly 1 Megabyte (1,000,000 bytes) on Windows, Android, and some other systems, it will show up as 976 Kilobytes, I am using a different base system.
This is the part that makes the joke functional, as you can actually download RAM, it just doesn't do anything. I was originally going to give it the RAM file extension, but I changed it, as VLC recognized it, and apparently RAM is an actual video format.